+++ lib/generators/happy_seed/omniauth/templates/app/controllers/omniauth_callbacks_controller.rb
@@ -1,18 +1,11 @@
 class OmniauthCallbacksController < Devise::OmniauthCallbacksController
-  def instagram
-    generic_callback( 'instagram' )
-  end
-  def facebook
-    generic_callback( 'facebook' )
-  end
-
-
   def generic_callback( provider )
-    @identity = Identity.find_for_oauth env["omniauth.auth"]
+    @identity = Identity.find_for_oauth request.env["omniauth.auth"]
 
     @user = @identity.user || current_user
+
     if @user.nil?
-      @user = User.create( email: @identity.email || "" )
+      @user = User.create( email: @identity.email, oauth_callback: true )
       @identity.update_attribute( :user_id, @user.id )
     end
 
@@ -22,9 +15,6 @@ class OmniauthCallbacksController < Devise::OmniauthCallbacksController
 
     if @user.persisted?
       @identity.update_attribute( :user_id, @user.id )
-      # This is because we've created the user manually, and Device expects a
-      # FormUser class (with the validations)
-      @user = FormUser.find @user.id
       sign_in_and_redirect @user, event: :authentication
       set_flash_message(:notice, :success, kind: provider.capitalize) if is_navigational_format?
     else