+++ lib/generators/happy_seed/omniauth/templates/spec/factories/oauth.rb
@@ -1,5 +1,7 @@
 FactoryGirl.define do
   factory :oauth_user, :class => 'User' do
-    
+    before(:create) do |user, evaluator|
+      user.oauth_callback = true
+    end
   end
 end