+++ lib/generators/happy_seed/omniauth/templates/user.rb
@@ -1,6 +1,6 @@
-class FormUser < User
+  attr_accessor :oauth_callback
   attr_accessor :current_password
-  
+    
   validates_presence_of   :email, if: :email_required?
   validates_uniqueness_of :email, allow_blank: true, if: :email_changed?
   validates_format_of     :email, with: Devise.email_regexp, allow_blank: true, if: :email_changed?
@@ -10,11 +10,10 @@ class FormUser < User
   validates_length_of       :password, within: Devise.password_length, allow_blank: true
 
   def password_required?
-    return false if email.blank?
+    return false if email.blank? || !email_required?
     !persisted? || !password.nil? || !password_confirmation.nil?
   end
 
   def email_required?
-    true
+    @oauth_callback != true
   end
-end